♻️ :: flatten param directory
