+++ domain/src/main/java/team/aliens/domain/_param/auth/SignInRequest.kt
@@ -1,0 +1,11 @@
+package team.aliens.domain._param.auth
+
+/**
+ * @author junsuPark
+ * A request, sent when when signing in into service
+ * @property accountId user id
+ */
+data class SignInRequest(
+    val accountId: String,
+    val password: String,
+)
