--- domain/src/main/java/team/aliens/domain/_param/auth/password/EditPasswordRequest.kt
@@ -1,12 +1,0 @@
-package team.aliens.domain._param.auth.password
-
-/**
- * @author junsuPark
- * A request used when editing password
- * @property currentPassword a current password
- * @property newPassword a new password to be set
- */
-data class EditPasswordRequest(
-    val currentPassword: String,
-    val newPassword: String,
-)
