+++ domain/src/main/java/team/aliens/domain/_param/file/UploadFileResponse.kt
@@ -1,0 +1,10 @@
+package team.aliens.domain._param.file
+
+/**
+ * @author junsuPark
+ * A response returned when uploading file succeed
+ * @property fileUrl an url of file, uploaded on server
+ */
+data class UploadFileResponse(
+    val fileUrl: String,
+)
