--- domain/src/main/java/team/aliens/domain/_param/file/uploadfile/UploadFileResponse.kt
@@ -1,10 +1,0 @@
-package team.aliens.domain._param.file.uploadfile
-
-/**
- * @author junsuPark
- * A response returned when uploading file succeed
- * @property fileUrl an url of file, uploaded on server
- */
-data class UploadFileResponse(
-    val fileUrl: String,
-)
