+++ domain/src/main/java/team/aliens/domain/_param/notice/QueryNoticesResponse.kt
@@ -1,0 +1,25 @@
+package team.aliens.domain._param.notice
+
+import java.util.*
+
+/**
+ * @author junsuPark
+ * A response returned when querying notices
+ * @property notices list of notices of the user
+ */
+data class QueryNoticesResponse(
+    val notices: List<NoticeInformation>,
+) {
+
+    /**
+     * @author junsuPark
+     * @property id id of unique notice
+     * @property title title of the notice
+     * @property createdAt created time of the notice
+     */
+    data class NoticeInformation(
+        val id: UUID,
+        val title: String,
+        val createdAt: String,
+    )
+}
