+++ domain/src/main/java/team/aliens/domain/_param/student/CheckStudentNumberResponse.kt
@@ -1,0 +1,10 @@
+package team.aliens.domain._param.student
+
+/**
+ * @author junsuPark
+ * A response returned when checking student's number, contains user's name
+ * @property name name of student, matching with student's number
+ */
+data class CheckStudentNumberResponse(
+    val name: String,
+)
