--- domain/src/main/java/team/aliens/domain/_param/student/checkstudentnumber/CheckStudentNumberResponse.kt
@@ -1,10 +1,0 @@
-package team.aliens.domain._param.student.checkstudentnumber
-
-/**
- * @author junsuPark
- * A response returned when checking student's number, contains user's name
- * @property name name of student, matching with student's number
- */
-data class CheckStudentNumberResponse(
-    val name: String,
-)
