--- domain/src/main/java/team/aliens/domain/_param/studyroom/querystudyroomapplytime/QueryStudyRoomApplyTimeResponse.kt
@@ -1,12 +1,0 @@
-package team.aliens.domain._param.studyroom.querystudyroomapplytime
-
-/**
- * @author junsuPark
- * A response returned when querying study room apply time
- * @property startAt the start time of applying study room
- * @property endAt the end time of applying study room
- */
-data class QueryStudyRoomApplyTimeResponse(
-    val startAt: String,
-    val endAt: String,
-)
