--- domain/src/main/java/team/aliens/domain/_param/user/editpassword/EditPasswordRequest.kt
@@ -1,12 +1,0 @@
-package team.aliens.domain._param.user.editpassword
-
-/**
- * @author junsuPark
- * A request sent when changing password without verification
- * @property currentPassword a password user using it now
- * @property newPassword a password to be set
- */
-data class EditPasswordRequest(
-    val currentPassword: String,
-    val newPassword: String,
-)
